Canonize Go translate snippets
