+++ translate/snippets/snippet_test.go
@@ -12,19 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package translate_snippets contains snippet code for the Translate API.
+// Package trsnip contains snippet code for the Translate API.
 // The code is not runnable.
-package translate_snippets
+package trsnip
 
 import (
 	"bytes"
 	"strings"
-
-	"golang.org/x/text/language"
-
 	"testing"
 
 	"github.com/GoogleCloudPlatform/golang-samples/internal/testutil"
+	"golang.org/x/text/language"
 )
 
 func TestTranslateText(t *testing.T) {
